requirements.txt
# Core Deep Learning
torch>=2.0.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
sounddevice>=0.4.6
opensmile>=2.5.0
scipy>=1.10.0

# Utilities
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0

# Real-time Communication
python-osc>=1.8.3

# Training & Logging (Optional)
tensorboard>=2.12.0
matplotlib>=3.7.0
